test.model.building_blocks.heads.test_detections_3_anchors: fix & clean
